refactor: rename package step3 for racingcar
