+++ src/main/kotlin/racingcar/CarFactory.kt
@@ -1,0 +1,6 @@
+package racingcar
+
+class CarFactory {
+    fun create(numberOfCars: Int): List<Car> =
+        List(numberOfCars) { Car() }
+}
