--- src/main/kotlin/step3/CarFactory.kt
@@ -1,13 +1,0 @@
-package step3
-
-class CarFactory {
-    fun create(numberOfCars: Int): List<Car> {
-        val cars: MutableList<Car> = mutableListOf()
-
-        for (i in 1..numberOfCars) {
-            cars.add(Car())
-        }
-
-        return cars
-    }
-}
