+++ src/test/kotlin/racingcar/CarRacingSpec.kt
@@ -1,10 +1,10 @@
-package step3
+package racingcar
 
 import io.kotest.core.spec.style.StringSpec
 import org.assertj.core.api.Assertions.assertThat
 
 class CarRacingSpec : StringSpec({
-    "자동차 경주" {
+    "자동차수와 움직임수를 지정하면 자동차 경주를 실행하여 결과를 반환한다" {
         // arrange
         val numberOfCars = 3
         val numberOfMoves = 5
