+++ src/test/kotlin/racingcar/CarSpec.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.ints.shouldBeInRange
